requirements.txt
# Core dependencies
telethon==1.34.0
python-dotenv==1.0.0
aiohttp==3.9.1

# Optional: for better async support
aiodns==3.1.1
charset-normalizer==3.3.2